refactoring...
* moved Mechanism -> sasl.Mechanism
* moved common test code -> testing package
* return sasl.Mechanism in scram constructor instead of *mechanism
